And yet another iteration of working on resources
