+++ pages/_app.jsx
@@ -5,6 +5,9 @@ import { useRouter } from 'next/router'
 const MyApp = ({ Component, pageProps }) => {
     return (
         <>
+        <div>
+            <h2>Header</h2>
+        </div>
             <Component {...pageProps} router={useRouter()} />
         </>
     )
